% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.count.R
\docType{package}
\name{coda.count-package}
\alias{coda.count}
\alias{coda.count-package}
\title{coda.count: Tools for Compositional Count Data Distributions}
\description{
coda.count contains functions to estimate and work with multivariate
count data models. The package contains functions to deal with the
Dirichlet-multinomial distributions and, specially, the logratio-normal-multinomial
distribution (lrnm), or its particular version, the logistic-normal-multinomial.

Different approaches to calculate or approximate the moments of a
lrnm are available: Montecarlo and quasi-Montecarlo
simulation, Gauss-Hermite quadrature and variational methods.

To approximate the lrnm distribution the package includes a conditional version
were certain counts can be fixed during the estimation.
}
\author{
\strong{Maintainer}: Marc Comas-Cufí \email{mcomas@imae.udg.edu}

}
