% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{ddm}
\alias{ddm}
\title{Probability mass function of a Dirichlet-multinomial distribution}
\usage{
ddm(x, alpha)
}
\arguments{
\item{x}{count vector}

\item{alpha}{Parameter alpha of a Dirichlet-Multinomial distribution}
}
\value{
probability mass function evaluated in x
}
\description{
Probability mass function of a Dirichlet-multinomial distribution
}
\examples{
X = apply(simplex_lattice(19,2), 2, rev)
sum(p <- apply(X, 1, ddm, c(1.5,2)))
names(p) = sprintf("(\%d,\%d)", X[,1], X[,2])
barplot(p, cex.axis = 0.8, cex.names = 0.8, las=2,
        main = paste('Dirichlet-multinomial',
                     'probability mass function', sep="\n"),
                     xlab = 'Count', ylab = 'Probability')
D = 3
X = simplex_lattice(10,3)
(pmf <- cbind(X, ddm(X, c(0.5,0.5,0.5))))
sum(pmf[,4])
}
