% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrnm.R
\name{dlrnm}
\alias{dlrnm}
\title{Probability mass function of a logratio-normal-multinomial distribution}
\usage{
dlrnm(x, mu, sigma, B = NULL, method = ifelse(length(mu) \%in\% 1:2,
  "hermite", "mc"), hermite.order = 100, hermite.step_by = 100,
  hermite.eps = 1e-06, hermite.max_steps = 10)
}
\arguments{
\item{x}{count vector}

\item{mu}{Parameter mu of a logratio-normal-multinomial distribution}

\item{sigma}{Parameter sigma of a logratio-normal-multinomial distribution}

\item{method}{Methods available: 'hermite' for Hermite integration, 'mc' for Monte Carlo integration}
}
\value{
probability mass function evaluated in x
}
\description{
Probability mass function of a logratio-normal-multinomial distribution
}
\examples{
X = apply(simplex_lattice(10,2), 2, rev)
sum(p <- apply(X, 1, dlrnm, -0.5, 0.1))
names(p) = sprintf("(\%d,\%d)", X[,1], X[,2])
barplot(p, cex.axis = 0.8, cex.names = 0.8)
}
