% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_lrnm.R
\name{fit_conditional_lrnm}
\alias{fit_conditional_lrnm}
\title{Estimate the parameters of a logratio-normal-multinomial distribution
when some logratios are fixed.}
\usage{
fit_conditional_lrnm(
  X,
  C = X > 0,
  B = NULL,
  method = "mc",
  low.dim = 0,
  hermite.order = 5,
  mc.nsim = 500,
  mc.znorm = NULL,
  eps = NULL,
  max_iter = 500
)
}
\arguments{
\item{X}{count sample}

\item{C}{conditioning variable TRUE variables are fixed}

\item{B}{clr-basis in which mu and sigma are interpreted with. Default basis is given by coda.base::ilr_basis(length(x))}

\item{method}{Method to use to estimate the parameters: 'hermite', 'laplace', 'mc' (default) or 'vem'}

\item{low.dim}{Number of dimensions used to calculate expected values, 0: all dimensions.}

\item{hermite.order}{order of Hermite polynomials}

\item{mc.nsim}{number of low-discrepancy vectors to be generated for
approximating during the E-step phase}

\item{mc.znorm}{matrix used to evaluate the montecarlo method, if not defined,
gaussian random variables generated with R are used.}

\item{eps}{precision used for the final estimates. 1e-5 for hermite method
and 1e-3 for montecarlo method.}

\item{max_iter}{maximum number of iterations for the iterative procedure
used to estimate the parameter}
}
\value{
Estimated parameters mu and sigma
}
\description{
Estimate the parameters of a logratio-normal-multinomial distribution
when some logratios are fixed.
}
