% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{fit_dm}
\alias{fit_dm}
\title{Estimate the parameters of a Dirichlet-multinomial distribution}
\usage{
fit_dm(X, eps = 1e-04, maxiter = 5000)
}
\arguments{
\item{X}{count sample}
}
\value{
Estimated parameters alpha
}
\description{
Estimate the parameters of a Dirichlet-multinomial distribution
}
\examples{
set.seed(1)
X = rdm(c(1,4,0.5), 100, n = 100)
fit_dm(X)
}
