% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_lrnm_low_dim.R
\name{fit_one_dimensional_conditional_lrnm}
\alias{fit_one_dimensional_conditional_lrnm}
\title{Estimate the parameters of a logratio-normal-multinomial distribution
when some logratios are fixed.}
\usage{
fit_one_dimensional_conditional_lrnm(
  X,
  C = X > 0,
  V = NULL,
  V.direction = "center",
  B = NULL,
  hermite.order = 5,
  eps = 1e-05,
  max_iter = 500,
  debug = FALSE
)
}
\arguments{
\item{X}{count sample}

\item{C}{conditioning variable TRUE variables are fixed}

\item{V}{a matrix (same size as X) containing the directions (log-contrast) used to find the expected values in the E-step for each observation.}

\item{V.direction}{'center' if sub-variety passing through the normal's center (default) or 'equal' using the equal weight for part.}

\item{B}{clr-basis in which mu and sigma are interpreted with. Default basis is given by coda.base::ilr_basis(length(x))}

\item{hermite.order}{order of Hermite polynomials}

\item{eps}{precision used for the final estimates. 1e-5 for Hermite method
and 1e-3 for Montecarlo method.}

\item{max_iter}{maximum number of iterations for the iterative procedure
used to estimate the parameter}

\item{debug}{if set to TRUE full list with debugging parameters is returned (this list is highly possible to change in different versions)}
}
\value{
Estimated parameters mu and sigma
}
\description{
Estimate the parameters of a logratio-normal-multinomial distribution
when some logratios are fixed.
}
