% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrnm.R
\name{lrnm_posterior_approx}
\alias{lrnm_posterior_approx}
\title{Estimate an approximation for the posterior distribution of the log-ratio-normal-multinomial
distribution.}
\usage{
lrnm_posterior_approx(X, mu, sigma, B, method = "laplace")
}
\arguments{
\item{X}{matrix of counts}

\item{mu}{mean parameter}

\item{sigma}{covariance parameter}

\item{B}{compositional basis}

\item{method}{Default 'laplace'}
}
\value{
A cube with the parameters
}
\description{
Estimate an approximation for the posterior distribution of the log-ratio-normal-multinomial
distribution.
}
