% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rdm}
\alias{rdm}
\title{Dirichlet-multinomial random sample}
\usage{
rdm(alpha, size, n = length(size), probs = FALSE)
}
\arguments{
\item{alpha}{Dirichlet-multinomial parameter}

\item{size}{vector to set the multinomial sampling size. vector is reused to have length equal parameter n}

\item{n}{sample size}

\item{probs}{logical indicating whether multinomial probabilities should be returned}
}
\value{
Dirichlet-multinomial random sample
}
\description{
Dirichlet-multinomial random sample
}
\examples{
rdm(c(1,1,1), 1000, n = 100)
}
