% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rlrnormal}
\alias{rlrnormal}
\title{logratio-normal random sample defined on the Simplex}
\usage{
rlrnormal(n, mu, sigma, B = coda.base::ilr_basis(length(mu) + 1))
}
\arguments{
\item{n}{sample size}

\item{mu}{logratio-normal-multinomial mu parameter}

\item{sigma}{logratio-normal-multinomial mu parameter}

\item{B}{basis in which the parameters are expressed with}
}
\value{
logratio-normal-multinomial random sample
}
\description{
logratio-normal random sample defined on the Simplex
}
\examples{
mu = c(0,1)
sigma = diag(2)
rlrnormal(mu, sigma, 10)
}
