% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rmultinomial}
\alias{rmultinomial}
\title{Multinomial random sample}
\usage{
rmultinomial(probs, size)
}
\arguments{
\item{probs}{matrix with number of rows equal to the number of samples to generate. Row number i contains the multinomial probability
to be  used when generating sample i.}

\item{size}{vector to set the multinomial sampling size. vector is reused to have length equal to the number of rows of probs}
}
\value{
multinomial random sample
}
\description{
Multinomial random sample
}
\examples{
probs = matrix(c(0.2, 0.4, 0.4,
                 0.5, 0.1, 0.4), byrow = TRUE, nrow = 6, ncol = 3)
size = c(10, 500, 10000)
rmultinomial(probs, size)
}
