% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplex_lattice}
\alias{simplex_lattice}
\title{Create all elements of an (n,D)-simplex lattice}
\usage{
simplex_lattice(n, D)
}
\arguments{
\item{n}{total sum}

\item{D}{simplex dimension}
}
\value{
matrix containing all elements of an (n,D)-simplex lattice
}
\description{
Create all elements of an (n,D)-simplex lattice
}
\examples{
simplex_lattice(4,3)
}
