% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.count.R
\docType{package}
\name{coda.count-package}
\alias{coda.count}
\alias{coda.count-package}
\title{coda.count: Tools for Compositional Count Data Distributions}
\description{
coda.count is a package to deal with a multivariate count dataset with the goal of using the dataset for a log-ratio analysis.
}
\author{
\strong{Maintainer}: Marc Comas-Cufí \email{mcomas@imae.udg.edu}

}
