% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{ddm}
\alias{ddm}
\title{Probability mass function of a Dirichlet-multinomial distributions}
\usage{
ddm(x, alpha)
}
\arguments{
\item{x}{count vector}

\item{alpha}{Parameter alpha of a Dirichlet-Multinomial distribution}
}
\value{
probability mass function evaluated in x
}
\description{
Probability mass function of a Dirichlet-multinomial distributions
}
\examples{
X = lattice_simplex(3,10)
(pmf <- cbind(X, apply(X, 1, ddm, c(0.5,0.5,0.5))))
sum(pmf[,4])
}
