% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrnm.R
\name{dlrnm}
\alias{dlrnm}
\title{Probability mass function of a logratio-normal-multinomial distribution}
\usage{
dlrnm(
  x,
  mu,
  sigma,
  B = NULL,
  method = ifelse(length(mu) \%in\% 1:4, "hermite", "mc"),
  hermite.order = 10,
  mc.nsim = 1000,
  mc.znorm = NULL
)
}
\arguments{
\item{x}{count vector}

\item{mu}{Parameter mu of a logratio-normal-multinomial distribution}

\item{sigma}{Parameter sigma of a logratio-normal-multinomial distribution}

\item{B}{clr-basis in which mu and sigma are interpreted with. Default basis is given by coda.base::ilr_basis(length(x))}

\item{method}{Methods available: 'hermite' for Hermite integration, 'mc' for Monte Carlo integration}

\item{hermite.order}{order of Hermite polynomials}

\item{mc.nsim}{number of low-discrepancy vectors to be generated for
approximating during the E-step phase}

\item{mc.znorm}{matrix used to evaluate the montecarlo method, if not defined,
gaussian random variables generated with R are used.}
}
\value{
probability mass function evaluated in x
}
\description{
Probability mass function of a logratio-normal-multinomial distribution
}
\examples{
X = apply(simplex_lattice(19,2), 2, rev)
sum(p <- dlrnm(X, -0.5, 0.1))
names(p) = sprintf("(\%d,\%d)", X[,1], X[,2])
barplot(p, cex.axis = 0.8, cex.names = 0.8, las=2,
        main = paste('Log-ratio-normal-multinomial',
                     'probability mass function', sep="\n"),
        xlab = 'Count', ylab = 'Probability')
}
