% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{fit_dm}
\alias{fit_dm}
\title{Estimate the parameters of a Dirichlet-multinomial distribution}
\usage{
fit_dm(X, eps = 1e-04, max_iter = 500)
}
\arguments{
\item{X}{count sample}

\item{eps}{precision used for the final estimates}

\item{max_iter}{maximum number of iterations for the iterative procedure used to estimate the parameter}
}
\value{
Estimated parameters alpha
}
\description{
Estimate the parameters of a Dirichlet-multinomial distribution
}
\examples{
set.seed(1)
X = rdm(n = 1000, size = 100, c(1,4,0.5))
fit_dm(X)
}
