% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrnm.R
\name{fit_lrnm}
\alias{fit_lrnm}
\title{Estimate the parameters of a logratio-normal-multinomial distribution}
\usage{
fit_lrnm(X, B = NULL, probs = FALSE, order = 5, eps = 1e-08,
  max_iter = 500)
}
\arguments{
\item{X}{count sample}
}
\value{
Estimated parameters mu and sigma
}
\description{
Estimate the parameters of a logratio-normal-multinomial distribution
}
