% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrnm.R
\name{fit_lrnm}
\alias{fit_lrnm}
\title{Estimate the parameters of a logratio-normal-multinomial distribution}
\usage{
fit_lrnm(X, B = NULL, probs = FALSE, method = "montecarlo",
  H.ini = NULL, montecarlo.n = 500, hermite.order = 5, Z = NULL,
  eps = NULL, max_iter = 500)
}
\arguments{
\item{X}{count sample}

\item{B}{clr-basis in which mu and sigma are interpreted with. Default basis is given by coda.base::ilr_basis(length(x))}

\item{probs}{boolean indicating if expected posterior probabilities are returned.}

\item{montecarlo.n}{number of samples in the Montecarlo integration process.}

\item{hermite.order}{order of Hermite polynomials}

\item{eps}{precision used for the final estimates. 1e-5 for hermite method and 1e-3 for montecarlo method.}

\item{max_iter}{maximum number of iterations for the iterative procedure used to estimate the parameter}
}
\value{
Estimated parameters mu and sigma
}
\description{
Estimate the parameters of a logratio-normal-multinomial distribution
}
