% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrnm.R
\name{fit_lrnm}
\alias{fit_lrnm}
\title{Estimate the parameters of a logratio-normal-multinomial distribution}
\usage{
fit_lrnm(
  X,
  B = NULL,
  probs = FALSE,
  method = "mc",
  hermite.order = 5,
  mc.nsim = 500,
  mc.znorm = NULL,
  eps = NULL,
  max_iter = 500
)
}
\arguments{
\item{X}{count sample}

\item{B}{clr-basis in which mu and sigma are interpreted with. Default basis is given by coda.base::ilr_basis(length(x))}

\item{probs}{boolean indicating if expected posterior probabilities are returned.}

\item{method}{Method to use to estimate the parameters: 'hermite', 'laplace', 'montecarlo' (default)}

\item{hermite.order}{order of Hermite polynomials}

\item{mc.nsim}{number of samples in the Montecarlo integration process.}

\item{eps}{precision used for the final estimates. 1e-5 for hermite method and 1e-3 for montecarlo method.}

\item{max_iter}{maximum number of iterations for the iterative procedure used to estimate the parameter}

\item{H.ini}{Coordinates used to initialise the expected posterior compositions given count data}

\item{Z}{matrix used to evaluate the montecarlo method, if not defined, gaussian random variables generated with R are used.}
}
\value{
Estimated parameters mu and sigma
}
\description{
Estimate the parameters of a logratio-normal-multinomial distribution
}
