% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rdm}
\alias{rdm}
\title{Dirichlet-multinomial random sample}
\usage{
rdm(n = NULL, size, alpha, probs = FALSE)
}
\arguments{
\item{n}{sample size}

\item{size}{vector to set the multinomial sampling size. vector is reused to have length equal parameter n}

\item{alpha}{Dirichlet-multinomial parameter}

\item{probs}{logical indicating whether multinomial probabilities should be returned}
}
\value{
Dirichlet-multinomial random sample
}
\description{
Dirichlet-multinomial random sample
}
\examples{
rdm(100, 1000, c(1,1,1))
rdm(size = c(1000, 100, 10, 2, 1), alpha = c(1,1,1))
}
