% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rlrnm}
\alias{rlrnm}
\title{logratio-normal-multinomial random sample}
\usage{
rlrnm(mu, sigma, size, n = length(size), probs = FALSE, B = NULL)
}
\arguments{
\item{mu}{logratio-normal-multinomial mu parameter}

\item{sigma}{logratio-normal-multinomial mu parameter}

\item{size}{vector to set the multinomial sampling size. vector is reused to have length equal parameter n}

\item{n}{sample size}

\item{probs}{logical indicating whether multinomial probabilities should be returned}
}
\value{
logratio-normal-multinomial random sample
}
\description{
logratio-normal-multinomial random sample
}
\examples{
mu = c(0,1)
sigma = diag(2)
rlrnm(mu, sigma, 1000, 10)
}
