% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rlrnormal}
\alias{rlrnormal}
\title{logratio-normal random sample defined on the Simplex}
\usage{
rlrnormal(n, mu, sigma, B = NULL)
}
\arguments{
\item{n}{sample size}

\item{mu}{logratio-normal-multinomial mu parameter}

\item{sigma}{logratio-normal-multinomial mu parameter}

\item{B}{basis in which the parameters are expressed with}
}
\value{
logratio-normal-multinomial random sample
}
\description{
logratio-normal random sample defined on the Simplex
}
\examples{
mu = c(0,1)
sigma = diag(2)
P.ilr <- rlrnormal(100, mu, sigma)
colMeans(coda.base::coordinates(P.ilr))
cov(coda.base::coordinates(P.ilr))

B = alr_basis(3)
P.alr <- rlrnormal(100, mu, sigma, B)
colMeans(coda.base::coordinates(P.alr, B))
cov(coda.base::coordinates(P.alr, B))
}
